Fix graphical bug caused by blacking out dungeon stairs
The stairs image (res344.png) in VDUNGEON.DAT includes two pixels of the left side of the door. Blacking out the stairs image entirely caused the start door to be partially erased because of this. This fix narrows the dungeon stairs image by two pixels and adjusts for this in the drawing code.