Increase the number of supported potion types to 32
This change makes 2 originally unused bits available for potion types, thereby increasing the maximum number of unique potions from 8 to 32.
The original game reserves 5 bits of the tile modifier byte for the frames of the potion bubble animation; the remaining (most significant) 3 bits are used for the potion types. However, the potion animation has only 7 frames so 3 bits is quite sufficient for the bubble animation.