Fix bug with loose tiles to the left of potions
A drawing bug occurs when a loose tile is placed to the left of a potion.
The right side of the loose tile is drawn as if it is an edge instead of part of a continuous floor.

This is an attempted fix. Unfortunately, the bug still occurs when the loose tile wobbles. I am not sure how this could be fixed.