Add scripting feature (using libtcc)
This adds scripting capabilities through the use of libtcc (part of TinyCC, a.k.a. tiny C compiler).
The basic idea is that a C script containing modifications is compiled at runtime, after which it is linked with the host process in memory (SDLPoP). SDLPoP and the compiled script can then access each other's functions and variables as necessary.
The capabilities are still quite limited, but it will be easy to expand the feature as we want in the future.

New files added:
* data/script/lib/libtcc1.a - part of libtcc; it is needed so that functions like printf can work in scripts
* data/script/pop_script.h - header file, to be #included in the scripts. It contains declarations for use in scripts
* script.h and script.c - new source files
* script.p1s - example script