Improved fix for "standing on thin air" bug
The changed fix no longer modifies the sequence table (as a workaround to prevent the kid falling off ledges, when standing up from crouching).
The old workaround worked by slightly moving the kid horizontally so falling off was not possible, but of course this horizontal movement is an unneeded deviation from the original game. The new method prevents the kid from falling in check_on_floor() in seg006.c.