Fix new falling bug caused by "glide through wall" fix
Char.fall_x should only be set to zero when facing toward a wall, not away from a wall.
With the "glide through wall" fix active, this could happen, because the game sometimes thinks the kid is "inside" a wall when he is actually just behind it. One reproducible example is the following set-up:

__
X __
_ XX
X XX   (underscore = floor, X = wall)

In the original game, if you drop down from the upper ledge, you should catch the lower ledge. But if Char.fall_x is set to zero, you miss the ledge.