Fix unintended sword strike immediately after drawing sword
Sometimes, the kid may automatically strike immediately after drawing the sword.
This especially happens when dropping down from a higher floor and then turning towards the opponent.

I think this happens because ctrl1_shift2 may still be -1 in draw_sword(), after which control_shift2 is set to -1 in rest_ctrl1().