Fix retreating out of a room without the room actually changing
By repeatedly pressing 'back' in a swordfight, you can retreat out of a room without the room changing. (Trick 35)

The game waits for a 'legal frame' (e.g. frame_170_stand_with_sword) until leaving is possible;
However, this frame is never reached if you press 'back' in the correct pattern!

Proposed solution: also allow the room to be changed on frame_157_walk_with_sword
Note that this means that the delay for leaving rooms in a swordfight becomes noticably shorter.