CusPop option: start in blind mode
This is slightly buggy (I am not sure why):
Bug: with start_in_blind_mode enabled, moving objects are not displayed until blind mode is toggled off+on??
Level 1 ending (e.g. by Shift+L) also seems to work around the issue.
The bug does not occur when is_blind_mode is only set to 1 after draw_level_first() has already been called once. But in that case the entire level is displayed (briefly), which is not the intended behavior...