Mod names as separate command line arguments
Changed usage of launching mods directly from the command line to `prince mod "Mod Name"
Because apparently `prince mod="Mod Name"` does not work well if there is whitespace in the mod name.
To prevent mod names conflicting with existing command line parameters, I made a change to check_param() so that it does not check 'sub-params' (e.g. the name of a mod supplied directly following a "mod" arg) and to make sure that it returns a pointer to the 'sub-arg' (handy for getting e.g. the mod name directly from check_param()).