Allow disabling the time limit completely
By setting `start_minutes_left` to -1 (or another negative number) the time limit can be disabled.
This also disables the in-game messages that report how much time is left every minute.
The elasped time is still kept track of, so that the shortest times will appear in the Hall of Fame (in ascending order, counting upwards from 0:00).
I changed the `rem_min` to a signed short integer to make this work.
(I wonder if `rem_min` was signed in the original DOS source... At least I suppose they did not count on large values, because they used a (signed) `%d` in the formatting string `%d MINUTES LEFT`)