Add SDLPoP splash screen as suggested by @EndeavourAccuracy
Info screen can be closed with a controller as well. I added a button state for the B button, otherwise that button would be the only button (out of A/B/X/Y) that is unable to close the splash screen, which would be strange...
Added SDLPOP_VERSION macro.
Updated ChangeLog.txt as well.