Enable launching the game using a custom executable (mod.dll)
This adds a way to let prince.exe act as a 'launcher' for SDLPoP mods.
The file `mod.dll` should be present in the mod folder (or, on Linux, `mod.so`).

I added configurations in CMakeLists.txt to help in building custom mod executables (use the target `mod` to build the mod). The macro `IS_SDLPOP_MOD_EXE` should be defined, only when building the mod. The `mod` target can be built either as a normal executable (easier to launch and test) or as a dll, depending on the value of `SHOULD_BUILD_MOD_AS_DLL`.

With the mod dll in place in the mod folder, you can launch the mod normally, through the command `prince mod <mod_name>` or by using the levelset option in SDLPoP.ini.

I tested it on Windows; I added Linux-specific code as well, but I haven't verified whether it works on Linux as well.

I restructured CMakeLists.txt a bit as well.
I haven't modified the Makefile or the other project files, I am unsure how to do that...

I did not notice any performance hits when loading a custom mod dll vs launching the custom exe directly. The only difference I could find is that prince.exe uses a few hundred kilobytes more memory.