Fix replays running out of sync
The replay 'special move' MOVE_EFFECT_END was added. In a replay, this special move will stop the feather effect / level 1 crouch immobilization (waiting for the music to stop playing is not reliable enough for a replay)
Also renamed replay_move to replay_move_type in replay.c to be more consistent with the other type names in the code base.