Replay: Add a unique identifier / class number for the replay format
This adds a way to tell apart different implementations/forks of SDLPoP in the replay format. If someone forks SDLPoP and modifies the format and/or version numbers of the replay format, there would be compatibility issues between the different forks. Ideally, the game should be able to recognize that a replay is "foreign" and not attempt to play it.

Also updated the outdated comments in `read_replay_header()`. And refactored away the `replay_header_type.magic` field because it is only used in `read_replay_header()`.