Fix warnings about redefinition of SDL_COMPILE_TIME_ASSERT
It seems that in SDL 2.0.6, this macro is no longer #undefined in the
SDL header files. Here, I've changed the name of the macro so it doesn't collide
anymore. (Alternative of course would be to #undef it manually or add an
ifdef guard, but I thought this might be the cleanest.)