In-game menu: toggle settings using keyboard or mouse
The settings menu can now be navigated using either the keyboard or mouse.
(Use the up/down arrow keys to select a setting, then left/right to turn that setting on or off.)
A number of toggles are working, e.g. sound on/off, enable 4:3 aspect ratio, enable lighting.
But not everything works yet.

Also:
* Removed some old unused code that I accidentally forgot to delete earlier.