Reduce CPU usage further (should be back to normal)
This adds delays in do_wait(), do_simple_wait() and in various cutscenes, in order to reduce CPU usage.
(And remove the delay in update_screen(), because that is no longer useful.)
The consequence of adding delays in do_wait() and do_simple_wait() is that timers will often slightly overshoot the desired duration, but we can compensate for that by 'backdating' timer_last_counter to one or several ticks earlier.
That way, the timer will 'catch up' and next frame will be drawn on time.
CPU usage should now be back to normal, because frames are drawn at the same rate as before.
The wipe transition is now too fast, unfortunately. We should probably set a fixed speed...