Fix cutscene left-to-right transition animation being too fast
This adds a delay after each frame, to slow down the animation to a fixed speed.
I set the animation speed at 180 frames per second for now, because this seemed to be a comfortable speed, and 180 fps is an integer number of times faster than the game's usual 60 fps.
But that number can easily be changed of course.