Add in-game menu
This adds an in-game menu.
The menu is shown when the game is paused, although that behavior can be changed back to the original pause behavior. Backspace also brings up the menu.
In-game settings are saved in a binary file SDLPoP.cfg. That file will be invalidated if you modify SDLPoP.ini, or if the prince executable changed (checked using CRC-32 hash).
Most options can now be changed while the game is running.
Added a way to easily disable and re-enable all customization options at once. (use_custom_options option in SDLPoP.ini)
The startup prompt to enable or disable fixes and enhancements was removed.

Miscellaneous other changes:
* Turning music off now works.
* Refactor: added a macro NAMES_LIST() for declaring and initializing name_list_type structs.
* Refactor: Simplified accessing cutscene function pointers by their index.