Fix hidden floors in level 12 not appearing until after the flashing stops
After uniting with the shadow in level 12, the hidden floors will not appear until after the flashing stops.
As mentioned by @EndeavourAccuracy here:
http://forum.princed.org/viewtopic.php?f=126&t=3512&p=22656#p22655

Also, fixed the 'move after sheathe' fix not being present in SDLPoP.ini and the menu. (I must have forgotten at the time... Sorry about that.) Changed the description for this fix as well, to be a bit more informative about what exactly is wrong.