Add most of the missing CusPop options
This adds most of the CusPoP options that were still missing.
For the entry pose (turning/falling/running) and seamless exit options: allowed them to be used in all levels.
Added key/value lists for option names, so that negative values can also be associated with a name. (for example, direction 'left' is -1)
Also fixed a few errors in the example mod.ini and in the settings menu.
The new options also get saved in replays, so I incremented the replay format version number.