Add option for controlling when the kid has the sword
This CusPop option was still missing.
It's similar to what CusPop allows you to do, but you can specify a specific level from which the kid will have the sword.
Added the named value 'Never' as an alias for 16 (higher than all level numbers), in order to be able to disable the sword completely.

And added the 'Never' alias to some of the other menu options as well, in case people want to e.g. completely disable a special event.
Also thought it might be a good idea to allow using 'never' in SDLPoP.ini as well, e.g., for the copyprot_level option (because using -1 there actually sets the value to UINT16_MAX...) Hence the changes in options.c.