Make hardware acceleration a runtime option
If we force hardware acceleration on a system where it's not available
SDL won't pick the software render driver as a fallback, resulting in a
black screen and an unplayable game (tested on WSLv1).
This commit lets SDL decide (it will prioritize a driver with hardware
acceleration anyway) which allows the game to run on platforms lacking
hardware acceleration without having to recompile, but let the user
enforce both options if they so wish.